Fully move files from «/» to canonical «/usr» locations
Instead of using the dh_movetousr hack, which will eventually disappear
and this work will need to be done anyway, do it right away. This means
all references use the canonical location, in the packaging, the binary
packages, and documentation, so that we avoid pointless indirections
from symlinks, and when confronted with such pathnames passing them to
stuff like «dpkg -S» will work properly.